+++ src/components/ProjectPageLink/styles/index.tsx
@@ -2,9 +2,12 @@ import { Link } from "react-router-dom";
 import styled from "styled-components/macro";
 
 export const StyledLink = styled(Link)`
-  font-size: 17px;
-  display: block;
+  font-size: 24px;
+  font-weight: 700;
+  font-family: poppins, sans-serif;
   color: inherit;
+  border-bottom: 1px solid ${({ theme }) => theme.pink};
+  align-self: center;
   img {
     margin-left: 10px !important;
   }
@@ -12,10 +15,4 @@ export const StyledLink = styled(Link)`
     color: inherit;
     text-decoration: underline;
   }
-`;
-
-export const Title = styled.div`
-  font-size: 24px;
-  font-weight: 700;
-  padding-bottom: 14px;
 `;
