+++ src/containers/ContentPage/ContentPage.tsx
@@ -13,104 +13,16 @@ import {
 import { TechContentType } from "../../data/content";
 import techContent from "../../data/content/index";
 import images from "../../images";
-import styled from "styled-components/macro";
+import {
+  Wrapper,
+  ContentWrapper,
+  ContentBody,
+  Title,
+  Divider,
+} from "./styles/";
 import ContentNavButton from "../../components/ComponentNavButton/ContentNavButton";
 import { getNextItem, getPrevItem } from "../../helpers/helpers";
-import ClientProjectLink from "../../components/ClinetProjectLink/ClientProjectLink";
-
-type WrapperProps = {
-  category: string;
-};
-
-const Wrapper = styled.div<WrapperProps>`
-  position: relative;
-  width: 100%;
-  display: flex;
-  padding-left: 90px;
-  margin: 50px 0 100px 0;
-  .background {
-    opacity: 0.12;
-    position: absolute;
-    top: -80px;
-    left: -300px;
-  }
-  img {
-    margin: auto;
-  }
-  @media screen and (max-width: 1000px) {
-    margin-top: 30px;
-    padding: 0 25px 0 25px;
-    img.background {
-      display: none;
-    }
-  }
-`;
-
-const ContentWrapper = styled.div`
-  text-align: left;
-  width: 1060px;
-  margin: auto;
-  position: relative;
-  :first-child {
-    display: inline-block;
-  }
-  svg.icon {
-    width: 80px;
-    height: 80px;
-  }
-  @media screen and (max-width: 1000px) {
-    width: 100%;
-    svg.icon {
-      width: 50px;
-      height: 50px;
-    }
-  }
-`;
-
-const ContentBody = styled.div`
-  width: 700px;
-  margin: 0 auto;
-  font-family: poppins, sans-serif;
-  .content-intro {
-    font-size: 24px;
-    margin-bottom: 34px;
-    a {
-      color: inherit;
-      border-bottom: 1px solid ${({ theme }) => theme.pink};
-      text-decoration: none;
-    }
-  }
-  @media screen and (max-width: 1000px) {
-    width: 100%;
-    .content-intro {
-      font-size: 18px;
-    }
-  }
-`;
-
-const Title = styled.div`
-  font-size: 120px;
-  font-weight: 700;
-  text-align: left;
-  display: inline-block;
-  vertical-align: middle;
-  padding-right: 50px;
-  @media screen and (max-width: 1000px) {
-    font-size: 52px;
-    padding-right: 30px;
-  }
-`;
-
-const Divider = styled.hr`
-  border: 0;
-  height: 1px;
-  background-image: linear-gradient(
-    to right,
-    rgba(255, 255, 255, 0),
-    rgba(255, 255, 255, 0.75),
-    rgba(255, 255, 255, 0)
-  );
-`;
+import ClientProjectLink from "../../components/ProjectPageLink/ProjectPageLink";
 
 const CategoryPage = () => {
   const { category, technology, setCategory, setTechnology } =
@@ -188,52 +100,59 @@ const CategoryPage = () => {
       />
       {content && (
         <ContentWrapper>
-          <div>
-            <BackLink category={category} />
-            <Title className={`title-${technology}`}>{technology}</Title>
-            <a href={content.docsLink} target="_blank" rel="noreferrer">
-              <svg className="icon" viewBox={"0 0 80 80"}>
-                <circle cx={40} cy={40} r={40} fill={"white"} />
-                <image x={15} y={15} href={imageLink} height={50} width={50} />
-              </svg>
-            </a>
-            <ContentBody>
-              <div
-                className="content-intro"
-                dangerouslySetInnerHTML={{ __html: content.intro }}
-              />
-              <SubContent contentData={content.content} />
-              {content.examples.length > 0 && (
-                <ExampleContent contentData={content.examples} />
-              )}
-
-              {content.reference.length > 0 && (
-                <ReferenceContent contentData={content.reference} />
-              )}
-
-              <ClientProjectLink
-                onClick={() => handleClick(category, technology)}
-              />
-
-              <Divider />
-              <ContentNavButton
-                onClick={() => setTechnology(previousTechnology)}
-                previousTechnology={previousTechnology}
-                nextTechnology={nextTechnology}
-                previousCategory={previousCategory}
-                nextCategory={nextCategory}
-                next={false}
-              />
-              <ContentNavButton
-                onClick={() => setTechnology(nextTechnology)}
-                previousTechnology={previousTechnology}
-                nextTechnology={nextTechnology}
-                previousCategory={previousCategory}
-                nextCategory={nextCategory}
-                next={true}
-              />
-            </ContentBody>
+          <BackLink category={category} />
+          <div className="content-head">
+            <div>
+              <Title className={`title-${technology}`}>{technology}</Title>
+              <a href={content.docsLink} target="_blank" rel="noreferrer">
+                <svg className="icon" viewBox={"0 0 80 80"}>
+                  <circle cx={40} cy={40} r={40} fill={"white"} />
+                  <image
+                    x={15}
+                    y={15}
+                    href={imageLink}
+                    height={50}
+                    width={50}
+                  />
+                </svg>
+              </a>
+            </div>
+            <ClientProjectLink
+              onClick={() => handleClick(category, technology)}
+            />
           </div>
+          <ContentBody>
+            <div
+              className="content-intro"
+              dangerouslySetInnerHTML={{ __html: content.intro }}
+            />
+            <SubContent contentData={content.content} />
+            {content.examples.length > 0 && (
+              <ExampleContent contentData={content.examples} />
+            )}
+
+            {content.reference.length > 0 && (
+              <ReferenceContent contentData={content.reference} />
+            )}
+
+            <Divider />
+            <ContentNavButton
+              onClick={() => setTechnology(previousTechnology)}
+              previousTechnology={previousTechnology}
+              nextTechnology={nextTechnology}
+              previousCategory={previousCategory}
+              nextCategory={nextCategory}
+              next={false}
+            />
+            <ContentNavButton
+              onClick={() => setTechnology(nextTechnology)}
+              previousTechnology={previousTechnology}
+              nextTechnology={nextTechnology}
+              previousCategory={previousCategory}
+              nextCategory={nextCategory}
+              next={true}
+            />
+          </ContentBody>
         </ContentWrapper>
       )}
     </Wrapper>
